tests: AssertEventually and NO_FATALS improvements
As implemented, the macro NO_FATALS() is a synonym for
ASSERT_NO_FATAL_FAILURES(). Unfortunately, ASSERT_NO_FATAL_FAILURES()
only checks for fatal failures in its argument expression, not in the
surrounding scope.

Also, for some reason, older versions of GCC don't seem to be able to
tolerate AssertEventually() wrapped by a NO_FATALS() call, which means
we have to check for timeout failures after AssertEventually() returns.

There are a few cases where we are using NO_FATALS() wrong so this patch
corrects those.

Included in this patch are the following changes:

1. Add a new macro called NO_PENDING_FATALS() that checks the
   surrounding scope for fatals.
2. Replace useless empty NO_FATALS() calls with NO_PENDING_FATALS().
3. Add new macro called ASSERT_EVENTUALLY() that wraps the 1-arg form of
   AssertEventually() and checks for failures. Use it wherever possible.
4. Add a call to NO_PENDING_FATALS() after use of the 2-arg form of
   AssertEventually().

Change-Id: Ia004214d21a09ca0523ea11d88d2cd4ff783a1ca
Reviewed-on: http://gerrit.cloudera.org:8080/6730
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins