Reduce arguments in file Read API
Because reads are now always read fully or result in an error if
all bytes can’t be read, we can simplify the Read API to pass
only a Slice instead of a slice, scratch, and length.

This patch changes the file Read APIs and all similar API calls
that call it.

Change-Id: I3e4df425a7387c28eb428e851a9f8c25167755d4
Reviewed-on: http://gerrit.cloudera.org:8080/6770
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins