[integration-tests] fixed flake in SecurityUnknownTskTest
As it turned out, running the ErrorUnavailableCommonOperations scenario
of the SecurityUnknownTskTest test along with other activity sometimes
leads to timeout on creating the test table: ~200 out of 1000 runs fail
if running with --stress-cpu-threads=8 (that was 0 out of 1000 if
running without the parallel stress treads).  Bumping up the timeout
from 1 to 3 seconds fixes the issue (0 out of 1000 runs fail even if
running with --stress-cpu-threads=10).

Change-Id: I57fbceacaaf20ccf242156cdea72afff1714e1c3
Reviewed-on: http://gerrit.cloudera.org:8080/6771
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>