Expose row format flags in KuduScanner
This adds a way to pass row format modifier flags to KuduScanner,
encoded in an int64_t as a bitset.

This API is marked as advanced and it's explicitely called
out that the user is responsible for knowing the row data format
that results from setting such flags and decoding the row data.

The only use for these flags, presently, is to set
PAD_UNIXTIME_MICROS_TO_16_BYTES, making sure that the server
pads slots for UNIXTIME_MICROS with an additional 8 bytes to
the left. In the future we might use these flags to (temporarily)
provide old/new row formats without having to change all the
row apis at once.

This adds a new test to all_types-itest that tests reading data
with the padding. The test scans a random projection to test
the padded column(s) in different positions.

Change-Id: I043b6514dc5fc307fc9c94eb41f3ae79796ba273
Reviewed-on: http://gerrit.cloudera.org:8080/6624
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins