Fix flaky test TestRecoverFromOpIdOverflow
This test is flaky because we race against the COMMIT message for the
first NO_OP in the WAL being written. It is currently hard to know when
the actual COMMIT message is written to the WAL so we use a workaround
to delete the first log segment before restarting the EMC in this test.

If the COMMIT doesn't get written in time then the tablet bootstrap
process doesn't prune that entry at startup time and it ends up in the
new log prior to the much higher-numbered log entries. The flaky test
failure was due to an out of order log index being detected, and looked
like the following error in the log:

F0504 21:28:21.128690 13908 raft_consensus_state.cc:502] Check failed: _s.ok() Bad status: Corruption: New operation's index does not follow the previous op's index. Current: 2147483648.2147483648. Previous: 1.1
*** Check failure stack trace: ***
    @     0x7fe0adef915d  google::LogMessage::Fail() at ??:0
    @     0x7fe0adefb05d  google::LogMessage::SendToLog() at ??:0
    @     0x7fe0adef8c99  google::LogMessage::Flush() at ??:0
    @     0x7fe0adefbaff  google::LogMessageFatal::~LogMessageFatal() at ??:0
    @     0x7fe0b57a8018  kudu::consensus::PendingRounds::AdvanceCommittedIndex() at ??:0
    @     0x7fe0b57848a5  kudu::consensus::RaftConsensus::NotifyCommitIndex() at ??:0
    @     0x7fe0b5749ccf  kudu::consensus::PeerMessageQueue::NotifyObserversOfCommitIndexChangeTask() at ??:0
    @     0x7fe0b575bdc1  kudu::internal::RunnableAdapter<>::Run() at ??:0

Change-Id: Ib382819307da04bb76d68d2c015dc0edd9f60267
Reviewed-on: http://gerrit.cloudera.org:8080/6808
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <davidralves@gmail.com>