[catalog_manager_tsk-itest] shortened test's run-time
The longer the test runs the more tables it creates and drops,
and that makes switching from one leader master to another longer.
The presence of some parallel activity exacerbates the problem because
re-elections happen more often.  That leads to timeouts on operations
performed by the client.

This patch shortens the test's run time to make it less flaky if running
on inferior VMs with other concurrent activity.

Change-Id: I21cca3d53d033f56f60e46fb5a3f67f5d7e44122
Reviewed-on: http://gerrit.cloudera.org:8080/6854
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <davidralves@gmail.com>