Improve ASSERT_STR_CONTAINS()
This can just defer to built-in gtest functions instead of being
constructed by us. The benefit is that the implementation doesn't have
to wrap the macro in a do ... while loop so we are able to use syntax
like:

  ASSERT_STR_CONTAINS(a, b) << extra_info;

whereas previously you couldn't append extra_info to the error message.

Also fix some uses of ASSERT_STR_CONTAINS that were missing semicolons.

Change-Id: Idea2abc0115e13dbb9f3853a5b821ccbf833d184
Reviewed-on: http://gerrit.cloudera.org:8080/6956
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>