KUDU-1294 fix issue with unreleased consumption
In DEBUG builds, the DCHECK assert on unreleased consumption fires
from time to time in the destructor of MemTracker objects.  The root
cause was found by Todd and described in the first comment on KUDU-1294.
This patch fixes the issue by re-ordering the sequence of removing the
transaction driver from the map and releasing its memory consumption.

As a test, I ran 1K+ iterations of security-faults-itest test built
in DEBUG configuration with --stress-cpu-threads=8.  Prior to this fix,
usually ~15 out of 1024 tests were failing due to the issue:
  http://dist-test.cloudera.org//job?job_id=aserbin.1494966417.6076
After the fix, there were not any failures in the whole 1K run:
  http://dist-test.cloudera.org//job?job_id=aserbin.1494963982.18117

Change-Id: I41d4fd4c0030af02ae5b1e83bd52833632ed6743
Reviewed-on: http://gerrit.cloudera.org:8080/6901
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins