locks: allow use of TSAN on many-core machines
On an 88-core machine, I found that TSAN builds would crash due to trying to
acquire more than 64 locks (a built-in maximum). Since percpu-rwlocks are an
optimization that isn't really relevant in TSAN, and in fact can make the
results a little harder to read, this changes percpu_rwlock in TSAN builds
to just use a simple lock.

I also fixed the OSX path to only use a single lock as well, rather than
allocate an array of locks. It was already just using a single one of
those locks (based on the address of the percpu_rwlock object), so
changing to only allocate a single one is simpler.

Change-Id: I08b41a5bbdbef9046b07d2908b2b824bb9dd2b76
Reviewed-on: http://gerrit.cloudera.org:8080/6905
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins