log: change default retention to one segment
We've been testing this on some various clusters recently and it doesn't
seem to have any adverse effects. Additionally, it reduces the amount
of space used by the WAL for idle tablets, and it reduces startup time
since there are fewer WALs to replay. So, it's probably a good idea to
change the default.

Change-Id: I3094f36949b3e518e872993b3026f8314e271a5e
Reviewed-on: http://gerrit.cloudera.org:8080/6907
Reviewed-by: David Ribeiro Alves <davidralves@gmail.com>
Tested-by: Kudu Jenkins