util: add fault injection of EIOs
This patch adds the functionality to inject EIOs in env_posix based on
flag-specified glob patterns indicating which paths should fail and a
flag-specified double indicating how often these failures should occur.

A similar flag inject_io_error is replaced by the latter flag, and
wrapper macros are updated to ensure the returned error expression is
only executed when necessary.

A test is added to env-test.cc to verify that EIOs are successfully
triggered by specifying the patterns.

Change-Id: I5190d5e7f296cf27b65f1ecd2c16f3315cc46a39
Reviewed-on: http://gerrit.cloudera.org:8080/6881
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <davidralves@gmail.com>