KUDU-1034 client does not failover due to timeout
This patch fixes the issue described by KUDU-1034: the client does not
mark the failed tablet server as 'failed' in case of timeout and
continues to use it over and over again to send further requests,
even if other tablet replicas might be available.

Besides the actual fix, this patch incorporates an integration test
(ClientFailoverTServerTimeoutITest.FailoverOnLeaderTimeout) written by Mike.

Change-Id: Icfcece485e4053d921ffdc865612b3e7b9a992a3
Reviewed-on: http://gerrit.cloudera.org:8080/6924
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins