Fix inaccuracy in LICENSE.txt regarding MIT license
LICENSE.txt incorrectly stated that the MIT license doesn't require
copyright notices to be propagated to binary distributions. That is true
only of the Boost license.

It is sort of a moot point because the only MIT license software that
was listed in the relevant section is a build-time-only dependency, and
thus doesn't need to be propagated up. But, better to be accurate so
that future additions aren't treated incorrectly.

Change-Id: I872a9b5845b761bdf6980e4d9eea68b2970ef9bc
Reviewed-on: http://gerrit.cloudera.org:8080/6929
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>