Fix flaky test TestRecoverFromOpIdOverflow (again)
The previous attempt to fix this in commit
f0580499dc50e8a47ff6251301cdc15b9b79edcb had a flaw, but this test
really does fix the primary source of the flakiness. What appears to
have happened in the previous attempt is the dist-test passed and then I
made a couple additional tweaks before committing it which actually
broke it again.

The only "real" code change (the aforementioned fix) is on lines
L367-L371, however while I was in this test I also "modernized" it a bit
by making it inherit from ExternalMiniClusterITestBase which resulted in
a net-negative line count in this patch.

I ran the current version of this patch on dist-test in DEBUG mode with
8 cpu stress threads, and 199/200 passed (there is a nearly 50% failure
rate with 8 stress threads without this fix). The one that failed
actually timed out (with no logs, so I have no idea what went wrong)
but it is likely some unrelated (infrastructure?) issue.

This was the dist-test job:

http://dist-test.cloudera.org/job?job_id=mpercy.1495321732.3266

Change-Id: I1f7326136479311ba2a84b384327e07d280df7c3
Reviewed-on: http://gerrit.cloudera.org:8080/6943
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>