KUDU-1911 improve missing required arg message
When using a tool that has a required positional argument, excluding that
argument or passing it as a flag caused a confusing error message to be
printed out ("must provide missing_argument"). This commit changes
the message to "must provide positional argument missing_argument" for
non-variadic arguments and "must provide variadic positional argument
missing_argument" for variadic arguments.

Change-Id: I2128a01da5c4929981b2ea46a32b0f7b9dd066e1
Reviewed-on: http://gerrit.cloudera.org:8080/6986
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Dan Burkert <danburkert@apache.org>